+++ src/engine/Collision/Detection/QuadTree.ts
@@ -239,12 +239,12 @@ export class QuadTree<TItem extends QuadTreeItem> {
   }
 
   debug(ctx: ExcaliburGraphicsContext) {
-    this.bounds.draw(ctx, { color: Color.Yellow });
+    this.bounds.debug(ctx, { color: Color.Yellow });
     if (this._isDivided) {
-      this.topLeft.bounds.draw(ctx, { color: Color.Yellow });
-      this.topRight.bounds.draw(ctx, { color: Color.Yellow });
-      this.bottomLeft.bounds.draw(ctx, { color: Color.Yellow });
-      this.bottomRight.bounds.draw(ctx, { color: Color.Yellow });
+      this.topLeft.bounds.debug(ctx, { color: Color.Yellow });
+      this.topRight.bounds.debug(ctx, { color: Color.Yellow });
+      this.bottomLeft.bounds.debug(ctx, { color: Color.Yellow });
+      this.bottomRight.bounds.debug(ctx, { color: Color.Yellow });
     }
   }
 }